requirements.txt
telethon
aiosqlite
python-dotenv
asyncio
